feat: 테이블 뷰 placeholder 생성 함수 extension으로 이동
UIControl, UITableView extension을 생성하여, 테이블 뷰에 데이터가 없을 경우 표시되는 빈화면을 생성하는 함수를 만들었습니다.
