+++ ReadingPiece/Controller/Library/Journal/JournalViewController.swift
@@ -50,7 +50,9 @@ extension JournalViewController: UITableViewDataSource, UITableViewDelegate {
         let count = Journal.dummyData.count
         if count == 0 {
             let message = "아직 인증이 없어요. \n매일 독서 시간과 소감을 기록하고 \n챌린지를 달성해요!"
-            self.setEmptyView(image: UIImage(named: "recordIcon")!, message: message)
+            tableView.setEmptyView(image: UIImage(named: "recordIcon")!, message: message, buttonTitle: "독서 시작하기") { [self] in
+                buttonAction()
+            }
         }
         return count
     }
@@ -179,52 +181,8 @@ extension JournalViewController: JournalOldestDelegate, JournalLatestDelegate {
 
 // 데이터가 없을 경우 표시되는 Placeholder
 extension JournalViewController {
-    // 테이블 뷰에 데이터가 없을 경우 표시되는 placeholder
-    func setEmptyView(image: UIImage, message: String) {
-        let emptyView = UIView(frame: CGRect(x: self.view.center.x, y: self.view.center.y, width: self.view.bounds.size.width, height: self.view.bounds.size.height))
-        let imageView = UIImageView()
-        let messageLabel = UILabel()
-        let button = UIButton()
-        imageView.translatesAutoresizingMaskIntoConstraints = false
-        messageLabel.translatesAutoresizingMaskIntoConstraints = false
-        button.translatesAutoresizingMaskIntoConstraints = false
-        
-        messageLabel.font = .NotoSans(.medium, size: 15)
-        messageLabel.textColor = .charcoal
-        button.titleLabel?.font = .NotoSans(.medium, size: 15)
-        button.makeRoundedButtnon("독서 시작하기", titleColor: .white, borderColor: UIColor.melon.cgColor, backgroundColor: .melon)
-        
-        emptyView.addSubview(imageView)
-        emptyView.addSubview(messageLabel)
-        emptyView.addSubview(button)
-        emptyView.backgroundColor = #colorLiteral(red: 0.9646214843, green: 0.9647600055, blue: 0.9645912051, alpha: 1)
-        
-        //imageView.centerYAnchor.constraint(equalTo: emptyView.centerYAnchor).isActive = true
-        imageView.topAnchor.constraint(equalTo: emptyView.topAnchor, constant: 42).isActive = true
-        imageView.centerXAnchor.constraint(equalTo: emptyView.centerXAnchor).isActive = true
-        imageView.widthAnchor.constraint(equalToConstant: 57).isActive = true
-        imageView.heightAnchor.constraint(equalToConstant: 63).isActive = true
-        messageLabel.topAnchor.constraint(equalTo: imageView.bottomAnchor, constant: 16).isActive = true
-        messageLabel.leftAnchor.constraint(equalTo: emptyView.leftAnchor, constant: 80).isActive = true
-        messageLabel.rightAnchor.constraint(equalTo: emptyView.rightAnchor, constant: -80).isActive = true
-        button.topAnchor.constraint(equalTo: messageLabel.bottomAnchor, constant: 40).isActive = true
-        button.leftAnchor.constraint(equalTo: emptyView.leftAnchor, constant: 40).isActive = true
-        button.rightAnchor.constraint(equalTo: emptyView.rightAnchor, constant: -40).isActive = true
-        button.heightAnchor.constraint(equalToConstant: 44).isActive = true
-        
-        button.addTarget(self, action: #selector(buttonAction(_:)), for: .touchUpInside)
-        
-        
-        imageView.image = image
-        messageLabel.text = message
-        messageLabel.numberOfLines = 0
-        messageLabel.textAlignment = .center
-        
-        self.tableView.backgroundView = emptyView
-        self.tableView.separatorStyle = .none
-    }
     
-    @objc func buttonAction (_ sender: UIButton!) {
+    func buttonAction () {
         print("독서 시작 - 홈탭으로 이동 후 타이머 VC로 이동해야 함")
         
         //let homeNavigationVC = MyNavViewController()
@@ -244,4 +202,6 @@ extension JournalViewController {
         self.navigationController?.setViewControllers(newViewControllers as! [UIViewController], animated: true)
         */
     }
+    
+    
 }
